api/requirements.txt
alembic==1.0.5
asn1crypto==0.24.0
bcrypt==3.1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudinary==1.16.0
cryptography==2.8
Flask==1.0.2
Flask-Classy==0.6.10
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jwcrypto==0.6.0
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==3.0.0rc2
mock==3.0.5
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
python-dateutil==2.7.5
python-editor==1.0.3
requests==2.21.0
simplejson==3.16.0
six==1.12.0
SQLAlchemy==1.2.15
urllib3==1.24.1
Werkzeug==0.14.1
factory_boy
googlemaps
awesome-slugify
python-dotenv

